+++ lib/promscrape/targetstatus.go
@@ -1,6 +1,7 @@
 package promscrape
 
 import (
+	"flag"
 	"fmt"
 	"io"
 	"sort"
@@ -11,6 +12,10 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/prompbmarshal"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/promrelabel"
 )
+
+var maxDroppedTargets = flag.Int("promscrape.maxDroppedTargets", 1000, "The maximum number of `droppedTargets` shown at /api/v1/targets page. "+
+	"Increase this value if your setup drops more scrape targets during relabeling and you need investigating labels for all the dropped targets. "+
+	"Note that the increased number of tracked dropped targets may result in increased memory usage")
 
 var tsmGlobal = newTargetStatusMap()
 
@@ -240,12 +245,18 @@ type droppedTarget struct {
 }
 
 func (dt *droppedTargets) Register(originalLabels []prompbmarshal.Label) {
+
 	key := promLabelsString(originalLabels)
 	currentTime := fasttime.UnixTimestamp()
 	dt.mu.Lock()
-	dt.m[key] = droppedTarget{
-		originalLabels: originalLabels,
-		deadline:       currentTime + 10*60,
+	if k, ok := dt.m[key]; ok {
+		k.deadline = currentTime + 10*60
+		dt.m[key] = k
+	} else if len(dt.m) < *maxDroppedTargets {
+		dt.m[key] = droppedTarget{
+			originalLabels: originalLabels,
+			deadline:       currentTime + 10*60,
+		}
 	}
 	if currentTime-dt.lastCleanupTime > 60 {
 		for k, v := range dt.m {